Show calendar list scrollbar
