+++ Calendr/Extensions/NSView.swift
@@ -7,14 +7,6 @@
 
 import Cocoa
 
-extension NSView {
-
-    func scrollTop() {
-        // NSTextView is flipped by default (which took me ages to figure out)
-        scroll(.init(x: 0, y: isFlipped ? 0 : frame.height))
-    }
-}
-
 // MARK: - Factory
 
 extension NSView {
